feat(board.ts): Change data type for grid public property
I realized that if I wanted to be able to determine who the winner is an
for the SoSGame, I couldn't have my Board class just accept one type of
data type, so I wanted to allow Board class to accept an object type or
an array. When I play the SoSGame, no only do I need to track what symbols
('S' or 'O') are in each cell, but I also need to track which Player placed
the symbols in each cell, because I misunderstood that SoS players had to
stick to 'S' or 'O' the entire game, but they can actually choose to place
ether an 'S' or an 'O' during their turn.