refactor(SoSBoard.tsx): Move placeSymbolInEmptyCell() from Onclick event to makeMove() SoSGame public method
I purposefully added the currentPlayerTurn parameter to the makeMove()
public method so there are no sneaky connections when it's explicitly clear
about which player is making the move. I felt the board cell onClick event
should be primarily concerned with putting the S/O in the cell to be displayed
and I made the makeMove() public method primarily concerned with setting
the actual cell value in the 2-D array that's stored in the board private
property.