feat(player.ts): Add sosCount private int property
This private property (sosCount) will be useful in determining the winner
of a simple game versus a general game. For the simple game, the player
whose sosCount reaches 1 first is the winner. For the general game, the
entire board must be filled with S/O and the player with the greater
sosCount is the winner.