fix(App.tsx): Make sure selectGameMode() rerenders my SoSBoard component
I was experiencing weird troubles where for some reason the turnCount on
my GeneralSoSGame was #2 for 2 consecutive times a symbol was placed in
an empty cell. I had to make sure that in addition to updating my sosGame
variable that I also rerender the SoSBoard component with the new value
sosGame that's passed to SoSBoard as a prop.