feat(sosGame.ts): Add latestValidMove private and public properties
The latestMoveSubject public property is used to gets access to a data
stream that updates with the latest value of the latestValidMove private
property. My RecordedSoSGame class that I will create will subscribe to
receive the latest value from the data stream from latestMoveSubject and
I can put all of my game recording code in the RecordedSoSGame class.