test(unit): Create test cases for recording a simple game between 2 human players
I just need to make sure that I can record the settings of the simple game
and the what moves by which players took place the entire game, so when
I want to replay back the game, I can read the text file line by line and
parse the long string I'm going to put all of the game moves in. The strings
that contain the recorded game settings and recorded turns will be comebined
to be put into a text file that the user can download.